% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agencias_de_bacia_sp}
\alias{agencias_de_bacia_sp}
\title{Agências de Bacia Hidrográfica no Estado de São Paulo}
\format{
Uma tibble com 4 linhas (cada uma representando uma agência) e 7 colunas, contendo as variáveis:
\describe{
  \item{nome}{Nome da agência}
  \item{sigla}{Sigla utilizada no site do SigRH para a agência}
  \item{sigla_agencia}{Sigla utilizada no site do SigRH para a agência,
   em letras minúsculas e sem acentuação (como usado nas urls de acesso ao site). }
  \item{sigla_ugrhi}{Sigla utilizada no site do SigRH para a UGRHI}
  \item{url_atas}{URL da página no SigRH onde essa agência disponibiliza atas}
  \item{site_sigrh}{URL da página da agência no SigRH}
  \item{site_alternativo}{URL da página da agência externa ao SigRH}
}
}
\source{
SigRH - \url{https://sigrh.sp.gov.br/}
}
\usage{
agencias_de_bacia_sp
}
\description{
Tabela agregada à partir de dados obtidos no SigRH \url{https://sigrh.sp.gov.br/}.
}
\examples{
agencias_de_bacia_sp

}
\keyword{datasets}
