% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comites_sp}
\alias{comites_sp}
\title{Comit\u00eas de Bacia Hidrogr\u00e1fica no Estado de S\u00e3o Paulo}
\format{
Uma tibble com 22 linhas (cada uma representando uma UGRHI) e 5 colunas, contendo as vari\u00e1veis:
\describe{
  \item{bacia_hidrografica}{Nome da Bacia Hidrogr\u00e1fica}
  \item{sigla_comite}{Sigla utilizada no site do SigRH para o Comit\u00ea}
  \item{n_ugrhi}{N\u00famero que representa a Unidade de Gerenciamento de Recursos H\u00eddricos \u00e0 qual o Comit\u00ea est\u00e1 relacionado}
  \item{numero_municipios}{N\u00famero de munic\u00edpios que pertencem \u00e0 Unidade de Gerenciamento de Recursos H\u00eddricos \u00e0 qual o Comit\u00ea est\u00e1 relacionado}
  \item{macrometropole_daee}{Vari\u00e1vel l\u00f3gica. Caso seja TRUE, este Comit\u00ea est\u00e1 parcialmente ou totalmente no territ\u00f3rio da Macrometr\u00f3pole Paulista, segundo a delimita\u00e7\u00e3o do DAEE ( \url{http://www.sigrh.sp.gov.br/}).}
}
}
\source{
SigRH - \url{http://www.sigrh.sp.gov.br/municipios}
}
\usage{
comites_sp
}
\description{
Tabela agregada \u00e0 partir de dados obtidos no SigRH \url{http://www.sigrh.sp.gov.br/municipios}.
}
\examples{
comites_sp

}
\keyword{datasets}
