% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_format.R
\name{get_file_format}
\alias{get_file_format}
\title{Extract file format}
\usage{
get_file_format(url)
}
\arguments{
\item{url}{character}
}
\value{
character
}
\description{
Extract the file format from a url. If is NA, returns a NA.
}
\examples{
file <- "https://github.com/hadley/pkg-dev/blob/master/1-intro.pdf"
get_file_format(file)
}
