% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status_code.R
\name{get_status_code}
\alias{get_status_code}
\title{Title}
\usage{
get_status_code(url)
}
\arguments{
\item{url}{character}
}
\value{

}
\description{
Title
}
\examples{
file <- "https://github.com/hadley/pkg-dev/blob/master/1-intro.pdf"
get_status_code(file)

}
