% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_obter_agenda.R
\name{obter_tabela_agenda_comites}
\alias{obter_tabela_agenda_comites}
\title{Obter tabelas sobre agenda dos Comitês de Bacia}
\usage{
obter_tabela_agenda_comites(
  sigla_do_comite = NULL,
  online = TRUE,
  path_arquivo = NULL
)
}
\arguments{
\item{sigla_do_comite}{Texto referente à sigla do comitê. É possível
# verificar na base:  \code{\link{comites_sp}}.}

\item{online}{Caso seja TRUE, a tabela será obtida consultando a página web.
Caso seja FALSE, a tabela será obtida usando um caminho, que deve ser
informado no argumento \code{path_arquivo}. O valor padrão é TRUE.}

\item{path_arquivo}{Caminho para o arquivo .html que será lido.
Isso só deve ser usado com o argumento online sendo FALSE.
O caminho para o arquivo deve ser o gerado pela função \code{\link{download_html}}.}
}
\value{
Retorna uma tibble.
}
\description{
Função para obter tabelas sobre agenda dos Comitês de Bacia no Estado de São Paulo
}
\examples{
obter_tabela_agenda_comites("at")
}
