% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raspar_pagina_sigrh.R
\name{raspar_pagina_sigrh}
\alias{raspar_pagina_sigrh}
\title{Raspar páginas do SigRH}
\usage{
raspar_pagina_sigrh(
  sigla_do_comite = NULL,
  conteudo_pagina = NULL,
  orgao = NULL,
  online = TRUE,
  path_arquivo = NULL
)
}
\arguments{
\item{sigla_do_comite}{Texto referente à sigla do comitê. É possível
# verificar na base:  \code{\link{comites_sp}}.}

\item{conteudo_pagina}{Qual é o tipo de página que deve ser raspada.
Atualmente aceita: atas, representantes, agenda.}

\item{orgao}{Qual é o órgão da página que deve ser raspada.
Atualmente aceita: cbh, agencia.}

\item{online}{Caso seja TRUE, a tabela será obtida consultando a página web.
Caso seja FALSE, a tabela será obtida usando um caminho, que deve ser
informado no argumento \code{path_arquivo}. O valor padrão é TRUE.}

\item{path_arquivo}{Caminho para o arquivo .html que será lido.
Isso só deve ser usado com o argumento online sendo FALSE.
O caminho para o arquivo deve ser o gerado pela função \code{\link{download_html}}.}
}
\value{
Retorna uma tibble.
}
\description{
Função para raspar páginas do SigRH
}
\examples{
raspar_pagina_sigrh("at", conteudo_pagina = "atas", orgao = "cbh")
}
