% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comites_sp}
\alias{comites_sp}
\title{Comitês de Bacia Hidrográfica no Estado de São Paulo}
\format{
Uma tibble.
}
\source{
SigRH - \url{http://www.sigrh.sp.gov.br/}
}
\usage{
comites_sp
}
\description{
MUDAR ISSO
}
\keyword{datasets}
