% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comites_sp}
\alias{comites_sp}
\title{Comitês de Bacia Hidrográfica no Estado de São Paulo}
\format{

}
\source{
SigRH - \url{http://www.sigrh.sp.gov.br/}
}
\usage{
comites_sp
}
\description{
MUDAR ISSO Base de dados contendo informações sobre os Sistemas
que abastecem a Região Metropolitana de São Paulo
}
\examples{
knitr::kable(comites_sp)

}
\keyword{datasets}
