% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_obter_atas.R
\name{obter_tabela_atas_comites}
\alias{obter_tabela_atas_comites}
\title{Função para obter tabelas de atas dos comitês}
\usage{
obter_tabela_atas_comites(n_comite)
}
\arguments{
\item{n_comite}{}
}
\value{
Retorna uma tibble.
}
\description{
Função para obter tabelas de atas dos comitês
}
\examples{
obter_tabela_atas_comites(3)
}
