% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_obter_atas.R
\name{obter_tabela_atas_comites}
\alias{obter_tabela_atas_comites}
\title{Obter tabelas sobre atas dos Comitês de Bacia}
\usage{
obter_tabela_atas_comites(sigla_do_comite)
}
\arguments{
\item{sigla_do_comite}{Texto referente à sigla do comitê. É possível verificar na base:  \code{\link{comites_sp}}.}
}
\value{
Retorna uma tibble. Uma base com dados coletados para todos os comitês está disponível em \code{\link{atas_comites}}.
}
\description{
Função para obter tabelas sobre atas dos Comitês de Bacia no Estado de São Paulo
}
\examples{
obter_tabela_atas_comites("at")
}
