% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_obter_representantes.R
\name{obter_tabela_representantes_comites}
\alias{obter_tabela_representantes_comites}
\title{Title}
\usage{
obter_tabela_representantes_comites(n_comite)
}
\arguments{
\item{n_comite}{}
}
\value{

}
\description{
Title
}
