% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{representantes_comites}
\alias{representantes_comites}
\title{Representantes dos Comit\u00eas de Bacia Hidrogr\u00e1fica no Estado de S\u00e3o Paulo}
\format{
Uma tibble, contendo:
\describe{
  \item{data_coleta_dados}{Data em que os dados foram coletados, no formato ano-m\u00eas-dia (yyyy-mm-dd)}
  \item{site_coleta}{Link do site onde os dados foram coletados}
  \item{comite}{Nome do comit\u00ea de Bacia Hidrogr\u00e1fica}
  \item{n_ugrhi}{N\u00famero que representa a Unidade de Gerenciamento de Recursos H\u00eddricos \u00e0 qual o Comit\u00ea est\u00e1 relacionado}
  \item{organizacao_representante}{Organiza\u00e7\u00e3o representante \u00e0 qual a pessoa na coluna 'nome' est\u00e1 vinculada, segundo informado pelo Comit\u00ea de Bacia}
  \item{nome}{Nome da pessoa representante, segundo informado pelo Comit\u00ea de Bacia}
  \item{email}{Email da pessoa representante, segundo informado pelo Comit\u00ea de Bacia}
  \item{cargo}{Cargo (Titular ou Suplente) da pessoa representante, segundo informado pelo Comit\u00ea de Bacia}
}
}
\source{
SigRH - \url{http://www.sigrh.sp.gov.br/}
}
\usage{
representantes_comites
}
\description{
Tabela contendo informa\u00e7\u00f5es sobre representantes dos Comit\u00eas de Bacia Hidrogr\u00e1fica no Estado de S\u00e3o Paulo
}
\examples{
head(representantes_comites)
}
\keyword{datasets}
