% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{representantes_comites}
\alias{representantes_comites}
\title{Representantes dos Comitês de Bacia Hidrográfica no Estado de São Paulo}
\format{
Uma tibble, contendo:
\describe{
  \item{data_coleta_dados}{Data em que os dados foram coletados, no formato ano-mês-dia (yyyy-mm-dd)}
  \item{comite}{Nome do comit}
}
}
\source{
SigRH - \url{http://www.sigrh.sp.gov.br/}
}
\usage{
representantes_comites
}
\description{
MUDAR ISSO
}
\examples{
head(representantes_comites)
}
\keyword{datasets}
